requirements.txt
streamlit
pyspark==3.4.0
delta-spark==2.4.0
pandas
